refactor: refactor package structure
