+++ src/main/java/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 import java.util.Arrays;
