+++ src/main/java/calculator/CalculatorRunner.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 import java.util.Arrays;
 import java.util.Scanner;
