+++ src/main/java/calculator/StringParser.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 import java.util.ArrayList;
 import java.util.List;
