+++ src/main/java/calculator/StringVerifier.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 public class StringVerifier {
