--- src/main/java/carracing/model/GamePlayerStrategy.java
@@ -1,5 +1,0 @@
-package carracing.model;
-
-public interface GamePlayerStrategy {
-    Integer generateScore();
-}
