+++ src/main/java/racing/controller/GameController.java
@@ -1,4 +1,4 @@
-package carracing.controller;
+package racing.controller;
 
 public interface GameController {
     public void run();
