+++ src/main/java/racing/domain/tactic/GameTacticStrategy.java
@@ -1,0 +1,5 @@
+package racing.domain.tactic;
+
+public interface GameTacticStrategy {
+    Integer generateScore();
+}
