+++ src/main/java/racing/view/InputView.java
@@ -1,6 +1,6 @@
-package carracing.view;
+package racing.view;
 
-import carracing.common.exception.InvalidInputFormatException;
+import racing.common.exception.InvalidInputFormatException;
 import java.util.List;
 import java.util.Scanner;
