+++ src/test/java/calculator/StringParserTest.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
