+++ src/test/java/calculator/StringVerifierTest.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
